build: Add optional Capella version image tag
For local development with models that have been migrated to multiple
Capella versions it is handy to build and maintain Docker images for
multiple Capella versions in parallel.

This commit makes it possible to execute
`CAPELLA_VERSION=5.2.0 DOCKER_TAG=5.2 make all` for instance.

In addition this commit refactors the make targets a bit and makes use
of GNU make's automatic variable substitution (here `$@` carrying the
target name). This enhances the readability and reduces risks for
editing mistakes.