feat: Add CLI images
Consider dedicated images (with and w/o T4C client) to serve a
containerised Capella that can be run via its command line interface
(CLI). This is possible using a virtual display.

One specific and actual use case is to supply a container with a custom
Capella bundle that can access a volume mounted Capella model to export
all representations (diagrams) as images.