feat: Distinguish between local or pipeline run
This is necessary because we cannot use localhost to access
the local git server when we are working in our pipeline. On the other
hand, we cannot use the IP address of the container when it is running
locally. Therefore, we add the ability to specify a Docker network other
than "host" as an environment variable. In this case, we then use the
container address and not localhost.