fix: Use `bash` as default shell in Makefile
We do only support `/bin/bash`, but it was not set as default.
If the SHELL was not set on the target, it has used the system shell
as fallback. This can also be `/bin/sh` or any other shell,
which we don't support. This can lead to errors when running Make
targets on these systems.

This commit sets `bash` as global default shell.