feat: Extend capella base by feature patches
This commit enables the addition of feature patches not only to
our T4C image but also to the Capella base image. We have
transformed the `patch_t4c_sh` it into a more general `patch.sh`
script, which is now integrated into the Capella base image.
To apply feature patches to images built on the Capella base, it's
necessary to ensure the availability of a directory containing the
feature patches zip files and `patch_info.csv`. This directory then
has to be provided in PATCH_DIR and then we only have to run patch.sh