fix: Extend environment instead of replacing it
Better keep the existing environment variables (e. g. PATH) and only
add `GIT_ASKPASS` if the according script is given.

If the parameter `env` passed to `subprocess.run` is not None, it mustbe
be a mapping that defines the environment variables for the new process;
these are used instead of the default behavior of inheriting the currently
process’ environment.

Locally I had the case that the script `/etc/git_askpass.py` was run
and failed with an error
`/usr/bin/env: 'python': No such file or directory`