fix: Ensure fresh and clean dirs for each run
Especially when run locally, the script must always empty the dirs that
play a role. Otherwise, Capella behaves differently between a first run
and a second run.

Also we write a workspace dir in /tmp instead of the current dir which
is `/opt/capella`. `/opt/capella` is a read-only dir and a non-techuser
cannot create or work in a workspace dir in this dir.