feat: Change password by directly writing to `/etc/shadow`
The use of `passwd` is not possible in restricted environments.
It produces `Authentication token manipulation error`.

In one of our environments, session creation stopped after an
OpenShift cluster update to version 4.11.
Our suspect is the change of the Pod security admission.
More information can be found in the release notes:
https://docs.openshift.com/container-platform/4.11/release_notes/ocp-4-11-release-notes.html#ocp-4-11-auth-pod-security-admission