feat: Switch to CLI for diagram cache update
This changes the diagram cache template so that it directly uses the
py-capellambse CLI to execute it. This makes it more robust, as a change
in py-capellambse does not affect the ci template. Additionally, a
separate `$RESOURCES` was previously required, which is now directly
included in `$ENTRYPOINT`, being consistent with what the py-capellambse
diagram cache expects.