feat: Remove `HTTP` related code
This commit removes any code and related documentation that used the
`HTTP` variables. This is possible because no T4C administrator
operation should be performed in both the backup and export scripts,
so we do not need to provide these variables. The fact that this is
possible was discovered in
https://github.com/DSD-DBS/capella-dockerimages/pull/249
and described in
https://github.com/DSD-DBS/capella-dockerimages/issues/253