feat: Open commit description window on save
This forces the description window to pop up automatically
when you save a change. This is done because it is quite
difficult to find this setting on your own, especially if you
do not know about it in the first place. In addition, the window still
offers the option to simply ignore providing a description,
so the previous default behavior is still included. However, it usually
makes sense to provide a description, so this is now much easier to do.