feat: Install `ssh` client in base image
At least one user uses `ssh-agent` forwarding to access ssh keys in the container.
An SSH client might be helful for other use-cases like cloning via SSH in general.

I raised the topic in the team and there were no bojections to add
the 20MB large installation of `openssh-client` to the image.