fix: Use build-arg for Capella download build step
Consider scope and usage of the `ARG` instruction in Dockerfiles.

The `ARG` instruction in a Dockerfile defines a variable that users can
pass at build-time to the builder with the `--build-arg` flag. This
variable is only available during the build process and is not available
in the final image unless explicitly passed to an `ENV` instruction.

When we define `ARG` before the `FROM` instruction, it is available for
use in the `FROM` instruction itself (for example, to specify a base
image dynamically). However, its value is not automatically carried over
to the build steps that follow unless you redefine it after the `FROM`
instruction.