fix: Fix an issue where modals could be unfocused causing soft-locks
By default, the Xpra web client allows users to unfocus modals. When a modal is triggered, the main
window is frozen though, so if the modal is unfocused users are left with a frozen main window. By
injecting some javascript, we can force modal windows to always be focused. This should reduce the
amount of users that accidentally trigger a modal and then think their connection froze.