fix: Disable overlay scrolling in Capella
Capella's diagram view doesn't add extra padding around the content.
With overlay scrolling enabled, this frequently leads to the scrollbars
popping up unintentionally, and the user being unable to properly
interact with items on the bottom or right edge of the view.

Disabling the overlay scrolling feature makes the scrollbars always
appear and reserve the space they need, thereby eliminating the issue.